Fixed protobuf-net error
Fixed up the issues with protobuf-net, added a way to add a serializer
and a packet to the default serialier. Users will call this method to
register their packet that use the default serializer or if they want
another serializer they can register it now to use it on a per packet
basis.