Fixed abhorrent client issue
Initial tests indicated I could trust the deconstructor to fire on a
field of a MonoBehaviour but it's possible I fudged that test now that I
think about it. This caused the network thread to spin during editor
mode and all other sorts of bad things. This is now banaided with a
crappy time based disconnection method. It may cause some GC issues, I'm
about to profile it as we speak.