Bug Fixes
Client: Race conditions still exist but reduced probability of null
reference exceptions. Fixed a MAJOR issue crashing the Unity editor.
This is a key deciding factor in the decision to become more test
oriented. This should have been caught.

Server: Fixed invalid cast exceptions when trying to get lists of the
currently connected clients and servers. Error on my part for thinking
that the runtime would check the implemented class type of an interface
to determine if it was a valid cast, it does not. This is now fixed.
You're now able to send messages out on the server shutdown; last moment
manual disconnecting of clients or etc can be done here.

Both: Fixed namespace issues. I'll be doing a namespace overhaul and
project/dll name overhaul soon. Collisions are bad I know.